requirements.txt
Django==2.2.17
django-adminactions==1.9
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-codemirror2==0.2
django-colorfield==0.4.1
django-js-asset==1.2.2
-e git+https://github.com/NuChwezi/django-singleton.git@7644a87f89eaa4d9265ee305dcac2902f04d7037#egg=django_singleton
Pillow==5.0.0
psycopg2==2.8.6
python-dateutil==2.7.0
pytz==2018.3
six==1.11.0
sqlparse==0.3.0
xlrd==1.2.0
xlwt==1.3.0
